+++ lib/page/task_list/task/task_edit_modal.dart
@@ -1,0 +1,64 @@
+ import 'package:flutter/material.dart';
+import 'package:imploop/domain/task.dart';
+import 'package:imploop/service/task_service.dart';
+
+class TaskEditModal extends StatelessWidget {
+  TaskEditModal({Key? key, required this.task}) : super(key: key);
+
+  static show(BuildContext context, Task task) {
+    return Navigator.push(
+      context,
+      MaterialPageRoute(
+        builder: (context) {
+          return TaskEditModal(task: task);
+        },
+        fullscreenDialog: true,
+      ),
+    );
+  }
+
+  final Task task;
+  final GlobalKey<FormFieldState<String>> nameKey =
+      GlobalKey<FormFieldState<String>>();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(),
+      body: Center(
+        child: Form(
+          child: Column(
+            children: [
+              TextFormField(
+                key: nameKey,
+                initialValue: task.name,
+              ),
+              ElevatedButton(
+                onPressed: () async {
+                  String? updatedName = nameKey.currentState != null
+                      ? nameKey.currentState!.value
+                      : null;
+                  if (updatedName != null) {
+                    if (await TaskService.editTask(task.taskId, updatedName)) {
+                      // Taskが追加されたことをスナックバーで通知
+                      ScaffoldMessenger.of(context).showSnackBar(
+                        const SnackBar(
+                          content: Text(
+                            "Taskが変更されました。",
+                          ),
+                        ),
+                      );
+                      // 前の画面に遷移
+                      Navigator.pop(context);
+                    }
+                  }
+                },
+                child: const Text('変更する'),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
